move buildTransaction function into Libify
