+++ src/components/TxBuilderResult.js
@@ -1,13 +1,5 @@
 import React from 'react';
 import {connect} from 'react-redux';
-import {
-  Account,
-  TransactionBuilder,
-  Operation,
-  Asset,
-  Memo,
-  UnsignedHyper,
-} from 'stellar-sdk';
 import {PubKeyPicker} from './FormComponents/PubKeyPicker';
 import {EasySelect} from './EasySelect';
 import Libify from '../utilities/Libify';
@@ -37,7 +29,7 @@ export default class TxBuilderResult extends React.Component {
       errorTitleText = 'Form validation errors:';
       finalResult = formatErrorList(validationErrors);
     } else {
-      let transactionBuild = buildTransaction(attributes, operations);
+      let transactionBuild = Libify.buildTransaction(attributes, operations);
 
       if (transactionBuild.errors.length > 0) {
         errorTitleText = `Transaction building errors:`;
@@ -82,53 +74,6 @@ function chooseState(state) {
   return {
     state: state.transactionBuilder,
   }
-}
-
-function buildTransaction(attributes, operations) {
-  let result = {
-    errors: [],
-    xdr: '',
-  };
-
-  try {
-    var account = new Account(attributes.sourceAccount, UnsignedHyper.fromString(attributes.sequence).subtract(1).toString());
-
-    let opts = {};
-    if (attributes.fee !== '') {
-      opts.fee = attributes.fee;
-    }
-
-    var transaction = new TransactionBuilder(account, opts)
-
-    if (attributes.memoType !== 'MEMO_NONE') {
-      try {
-        transaction = transaction.addMemo(Libify.Memo({
-          type: attributes.memoType,
-          content: attributes.memoContent,
-        }));
-      } catch(e) {
-        console.error(e);
-        result.errors.push(`Memo: ${e.message}`);
-      }
-    }
-
-    _.each(operations, (op, index) => {
-      try {
-        transaction = transaction.addOperation(Libify.Operation(op.name, op.attributes));
-      } catch(e) {
-        console.error(e);
-        result.errors.push(`Operation #${index + 1}: ${e.message}`);
-      }
-    })
-
-    transaction = transaction.build();
-    result.xdr = transaction.toEnvelope().toXDR('base64');
-  } catch(e) {
-    console.error(e);
-    result.errors.push(e.message);
-  }
-
-  return result;
 }
 
 function formatErrorList(errors) {
